requirements.txt
django~=4.2.6
flake8~=6.1.0

pyjwt~=2.8.0
requests~=2.31.0
django_apscheduler~=0.6.2

uwsgi
mysqlclient

openai~=1.2.3
numpy

backoff

faiss-cpu

matplotlib